migrated all reductions-related tests to test_reductions.py
